+++ plone/app/event/at/content.py
@@ -29,6 +29,7 @@ from plone.formwidget.recurrence.at.widget import RecurrenceWidget
 from plone.uuid.interfaces import IUUID
 from zope.component import adapts
 from zope.event import notify
+from zope.globalrequest import getRequest
 from zope.interface import implements
 from zope.lifecycleevent import ObjectModifiedEvent
 
@@ -279,7 +280,6 @@ ATEventSchema = ATContentTypeSchema.copy() + atapi.Schema((
 
     atapi.StringField(
         'eventUid',
-        accessor='event_uri',
         write_permission=ModifyPortalContent,
         widget=atapi.StringWidget(
             visible={'edit': 'invisible', 'view': 'invisible'}),
@@ -643,6 +643,16 @@ def data_postprocessing(obj, event):
     start_field.set(obj, start.toZone('UTC'))
     end_field.set(obj, end.toZone('UTC'))
 
+    if not obj.getEventUid():
+        # event_uid has to be set for icalendar data exchange.
+        uid = IUUID(obj)
+        request = getRequest()
+        domain = request.get('HTTP_HOST')
+        obj.setEventUid('%s%s' % (
+            uid,
+            domain and '@%s' % domain or ''
+        ))
+
     obj.reindexObject()
 
 
@@ -808,7 +818,7 @@ class EventAccessor(object):
 
     @property
     def event_uid(self):
-        return self.context.event_uid()
+        return self.context.getEventUid()
     @event_uid.setter
     def event_uid(self, value):
         self.context.setEventUid(value)