+++ plone/app/event/dx/interfaces.py
@@ -26,8 +26,3 @@ class IDXEventAttendees(Interface):
 class IDXEventContact(Interface):
     """ Marker interface for Dexterity events with contact information.
     """
-
-
-class IDXEventUID(Interface):
-    """ Marker interface for Dexterity events with UID, for ical sync.
-    """