+++ CHANGES.rst
@@ -12,6 +12,10 @@ Changelog
     The plone.app.event.dx.event type and plone.app.event.dx:default profile
     are deprecated and will be removed in a future version of plone.app.event.
 
+- Add event_uid field to ATEvent and IEventBasic and an index, which is used
+  for icalendar data exchange and synchronisation.
+  [cillianderoiste, thet]
+
 - Don't show the repeat forever button in the recurrence widget.
   [thet]